+++ src/apps/cc/CMakeLists.txt
@@ -25,7 +25,7 @@ set(
     detector/linker/pot.cpp
     detector/linker/strategy.cpp
     detector/linker.cpp
-    detector/templatewaveformprocessor.cpp
+    detector/template_waveform_processor.cpp
     eventstore.cpp
     exception.cpp
     filter.cpp
