+++ src/apps/cc/detector/linker.h
@@ -14,7 +14,7 @@
 #include "linker/association.h"
 #include "linker/pot.h"
 #include "linker/strategy.h"
-#include "templatewaveformprocessor.h"
+#include "template_waveform_processor.h"
 
 namespace Seiscomp {
 namespace detect {
