+++ src/apps/cc/detector/template_waveform_processor.h
